requirements.txt
# requirements.txt

aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
docstring_parser==0.16
filelock==3.16.1
fire==0.7.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langsmith==0.1.140
MarkupSafe==3.0.2
mistral_common==1.4.4
mistral_inference==1.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.11
packaging==24.1
pillow==10.4.0
propcache==0.2.0
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.21.0
safetensors==0.4.5
sentencepiece==0.2.0
simple-parsing==0.1.6
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.6
transformers==4.46.2
typing_extensions==4.12.2
urllib3==2.2.3
xformers==0.0.28.post3
yarl==1.17.1
